Moved the specific upload functions into the general utils.py file
